chore(type): Move types files into src directory
