+++ src/repositories/articleRepository.ts
@@ -1,6 +1,6 @@
 import { Article, ArticleComment } from '@prisma/client';
 import { prisma } from '../config/prismaClient.js';
-import { CreateArticleCommentDto, CreateArticleDto, UpdateArticleCommentDto, UpdateArticleDto } from '../../types/article.js';
+import { CreateArticleCommentDto, CreateArticleDto, UpdateArticleCommentDto, UpdateArticleDto } from '../types/article.js';
 
 class articleRepository {
     create = async (ArticleDto: CreateArticleDto) => {
