+++ src/types/article.d.ts
